feat: change line_state to enum
