+++ src/domain/Profile.ts
@@ -1,24 +1,24 @@
-import Line, { LineResponse } from './Line';
-import ForwardOption, { FORWARD_KEYS } from './ForwardOption';
-import newFrom from '../utils/new-from';
-import SipLine from './SipLine';
-import Incall, { GenericIncall } from './Incall';
+import Line, { LineResponse } from "./Line";
+import ForwardOption, { FORWARD_KEYS } from "./ForwardOption";
+import newFrom from "../utils/new-from";
+import SipLine from "./SipLine";
+import Incall, { GenericIncall } from "./Incall";
 
 export const STATE = {
-  AVAILABLE: 'available',
-  UNAVAILABLE: 'unavailable',
-  INVISIBLE: 'invisible',
-  DISCONNECTED: 'disconnected',
-  AWAY: 'away',
+  AVAILABLE: "available",
+  UNAVAILABLE: "unavailable",
+  INVISIBLE: "invisible",
+  DISCONNECTED: "disconnected",
+  AWAY: "away",
 };
-export const LINE_STATE = {
-  AVAILABLE: 'available',
-  HOLDING: 'holding',
-  RINGING: 'ringing',
-  TALKING: 'talking',
-  UNAVAILABLE: 'unavailable',
-  PROGRESSING: 'progressing',
-};
+export enum LINE_STATE {
+  AVAILABLE = "available",
+  HOLDING = "holding",
+  RINGING = "ringing",
+  TALKING = "talking",
+  UNAVAILABLE = "unavailable",
+  PROGRESSING = "progressing",
+}
 
 type ProfileResponse = {
   groups: Array<{
@@ -141,48 +141,64 @@ export default class Profile {
 
   ringSeconds: number | null | undefined;
 
-  voicemail: {
-    id: number;
-    name: string;
-  } | null | undefined;
+  voicemail:
+    | {
+        id: number;
+        name: string;
+      }
+    | null
+    | undefined;
 
   status: string;
 
   subscriptionType: number | null | undefined;
 
-  agent: {
-    firstname: string;
-    id: number;
-    lastname: string;
-    number: string;
-  } | null | undefined;
+  agent:
+    | {
+        firstname: string;
+        id: number;
+        lastname: string;
+        number: string;
+      }
+    | null
+    | undefined;
 
   switchboards: Array<any>;
 
-  callPickupTargetUsers: Array<{
-    firstname: string;
-    lastname: string;
-    uuid: string;
-  }> | null | undefined;
+  callPickupTargetUsers:
+    | Array<{
+        firstname: string;
+        lastname: string;
+        uuid: string;
+      }>
+    | null
+    | undefined;
 
   static parse(plain: ProfileResponse): Profile {
     return new Profile({
       id: plain.uuid,
-      firstName: plain.firstName || plain.firstname || '',
-      lastName: plain.lastName || plain.lastname || '',
+      firstName: plain.firstName || plain.firstname || "",
+      lastName: plain.lastName || plain.lastname || "",
       email: plain.email,
-      lines: plain.lines.map(line => Line.parse(line)),
+      lines: plain.lines.map((line) => Line.parse(line)),
       incalls: plain.incalls,
       username: plain.username,
-      mobileNumber: plain.mobile_phone_number || '',
+      mobileNumber: plain.mobile_phone_number || "",
       ringSeconds: plain.ring_seconds,
-      forwards: [ForwardOption.parse(plain.forwards.unconditional, FORWARD_KEYS.UNCONDITIONAL), ForwardOption.parse(plain.forwards.noanswer, FORWARD_KEYS.NO_ANSWER), ForwardOption.parse(plain.forwards.busy, FORWARD_KEYS.BUSY)],
+      forwards: [
+        ForwardOption.parse(
+          plain.forwards.unconditional,
+          FORWARD_KEYS.UNCONDITIONAL
+        ),
+        ForwardOption.parse(plain.forwards.noanswer, FORWARD_KEYS.NO_ANSWER),
+        ForwardOption.parse(plain.forwards.busy, FORWARD_KEYS.BUSY),
+      ],
       doNotDisturb: plain.services.dnd.enabled,
       subscriptionType: plain.subscription_type,
       voicemail: plain.voicemail,
       switchboards: plain.switchboards || [],
       agent: plain.agent,
-      status: '',
+      status: "",
       callPickupTargetUsers: plain.call_pickup_target_users || [],
       onlineCallRecordEnabled: plain.online_call_record_enabled,
     });
@@ -270,7 +286,9 @@ export default class Profile {
 
   setForwardOption(forwardOption: ForwardOption) {
     const updatedForwardOptions = this.forwards.slice();
-    const index = updatedForwardOptions.findIndex(forward => forward.is(forwardOption));
+    const index = updatedForwardOptions.findIndex((forward) =>
+      forward.is(forwardOption)
+    );
     updatedForwardOptions.splice(index, 1, forwardOption);
     this.forwards = updatedForwardOptions;
     return this;
@@ -285,5 +303,4 @@ export default class Profile {
     this.state = state;
     return this;
   }
-
 }
